Restructure Fuzzer Seed Data
**Merging this change requires a corresponding PR to be merged at the same time
in the GitPython repo to update the OSS-Fuzz scripts to prevent breaking the
builds.**

The most significant change introduced here is replacing the `.zip` files
containing seed corpora with sub-directories for each fuzz target containing an
uncompressed corpus. This change makes it easier to update each corpus on a
per-input-blob basis, as well as making the content of each corpus easier to
inspect manually.

The inputs added in this commit are taken directly from the corpora backups
generated by ClusterFuzz/OSS-Fuzz and can (and should) likely be reduced to
their minimal sets in a follow-up PR.

Other changes:
- Renamed the `dict` sub-directory to `dictionaries` for clarity.
- Updated the README to better document what this repository contains. The README
  could use some improvement as well, including instructions for contributing and
  generating new inputs, but for now it should be good enough (I hope.) The
  "Dictionaries" section was migrated from the GitPython fuzzing README.